# Move serialization of .contract_class.json artifact to separate thread