--- a/scarb/src/compiler/compilers/starknet_contract/artifacts_writer.rs
+++ b/scarb/src/compiler/compilers/starknet_contract/artifacts_writer.rs
@@ -3,45 +3,22 @@ use crate::compiler::compilers::starknet_contract::{ContractFileStemCalculator, 
 use crate::compiler::helpers::write_json_with_byte_count;
 use crate::core::{PackageName, Workspace};
 use crate::flock::Filesystem;
+use crate::internal::artifacts_writer::{File, Request};
 use cairo_lang_compiler::db::RootDatabase;
 use cairo_lang_defs::ids::NamedLanguageElementId;
 use cairo_lang_starknet::contract::ContractDeclaration;
-use cairo_lang_starknet_classes::abi::Contract;
 use cairo_lang_starknet_classes::casm_contract_class::CasmContractClass;
-use cairo_lang_starknet_classes::contract_class::{ContractClass, ContractEntryPoints};
-use cairo_lang_utils::bigint::BigUintAsHex;
+use cairo_lang_starknet_classes::contract_class::ContractClass;
 use indoc::formatdoc;
 use itertools::{Itertools, izip};
 use scarb_stable_hash::short_hash;
 use serde::Serialize;
 use smol_str::SmolStr;
+use std::sync::{Arc, mpsc};
 use tracing::trace_span;
 
-const MAX_SIERRA_PROGRAM_FELTS: usize = 81290;
 const MAX_CASM_PROGRAM_FELTS: usize = 81290;
-const MAX_CONTRACT_CLASS_BYTES: usize = 4089446;
 const MAX_COMPILED_CONTRACT_CLASS_BYTES: usize = 4089446;
-
-// Represents a contract in the Starknet network as defined in Starknet JSON-RPC spec:
-// https://github.com/starkware-libs/starknet-specs/blob/2030a650be4e40cfa34d5051a0334f375384a421/api/starknet_api_openrpc.json#L3030
-#[derive(Clone, Debug, Serialize)]
-struct ContractClassNoDebug<'a> {
-    sierra_program: &'a Vec<BigUintAsHex>,
-    contract_class_version: &'a String,
-    entry_points_by_type: &'a ContractEntryPoints,
-    abi: &'a Option<Contract>,
-}
-
-impl<'a> ContractClassNoDebug<'a> {
-    fn new(contract_class: &'a ContractClass) -> Self {
-        Self {
-            sierra_program: &contract_class.sierra_program,
-            contract_class_version: &contract_class.contract_class_version,
-            entry_points_by_type: &contract_class.entry_points_by_type,
-            abi: &contract_class.abi,
-        }
-    }
-}
 
 #[derive(Debug, Serialize)]
 struct StarknetArtifacts {
@@ -107,16 +84,23 @@ pub struct ArtifactsWriter {
     target_dir: Filesystem,
     target_name: SmolStr,
     extension_prefix: Option<String>,
+    artifacts_writer: mpsc::Sender<Request>,
 }
 
 impl ArtifactsWriter {
-    pub fn new(target_name: SmolStr, target_dir: Filesystem, props: Props) -> Self {
+    pub fn new(
+        target_name: SmolStr,
+        target_dir: Filesystem,
+        props: Props,
+        artifacts_writer: mpsc::Sender<Request>,
+    ) -> Self {
         Self {
             sierra: props.sierra,
             casm: props.casm,
             target_dir,
             target_name,
             extension_prefix: None,
+            artifacts_writer,
         }
     }
 
@@ -131,7 +115,7 @@ impl ArtifactsWriter {
         self,
         contract_paths: Vec<String>,
         contracts: &Vec<ContractDeclaration>,
-        classes: &[ContractClass],
+        classes: Vec<ContractClass>,
         casm_classes: &[Option<CasmContractClass>],
         db: &mut RootDatabase,
         ws: &Workspace<'_>,
@@ -146,7 +130,8 @@ impl ArtifactsWriter {
             .map(|ext| format!(".{ext}"))
             .unwrap_or_default();
 
-        for (declaration, class, casm_class) in izip!(contracts, classes, casm_classes) {
+        for (declaration, class, casm_class) in izip!(contracts, classes.into_iter(), casm_classes)
+        {
             let contract_name = declaration.submodule_id.name(db);
             let contract_path = declaration.module_id().full_path(db);
 
@@ -164,36 +149,20 @@ impl ArtifactsWriter {
             );
 
             if self.sierra {
-                let sierra_felts = class.sierra_program.len();
-                if sierra_felts > MAX_SIERRA_PROGRAM_FELTS {
-                    ws.config().ui().warn(formatdoc! {r#"
-                        Sierra program exceeds maximum byte-code size on Starknet for contract `{}`:
-                        {MAX_SIERRA_PROGRAM_FELTS} felts allowed. Actual size: {sierra_felts} felts.
-                    "#, contract_stem.clone()});
-                }
-
                 let file_name = format!("{file_stem}{extension_prefix}.contract_class.json");
 
-                let class_size = write_json_with_byte_count(
-                    &file_name,
-                    "output file",
-                    &self.target_dir,
-                    ws,
-                    class,
-                )?;
-                if class_size > MAX_CONTRACT_CLASS_BYTES {
-                    // Debug info is omitted on Starknet.
-                    // Only warn if size without debug info exceeds the limit as well.
-                    let rpc_class = ContractClassNoDebug::new(class);
-                    let rpc_class_size = serde_json::to_vec(&rpc_class)?.len();
+                self.artifacts_writer
+                    .send(Request::ContractClassArtifact {
+                        file: File {
+                            file_name: file_name.clone(),
+                            description: "output file".to_string(),
+                            target_dir: self.target_dir.clone(),
+                        },
+                        value: Arc::new(class),
+                        contract_stem: contract_stem.clone(),
+                    })
+                    .expect("failed to send program artifact request");
 
-                    if rpc_class_size > MAX_CONTRACT_CLASS_BYTES {
-                        ws.config().ui().warn(formatdoc! {r#"
-                            Contract class size exceeds maximum allowed size on Starknet for contract `{}`:
-                            {MAX_CONTRACT_CLASS_BYTES} bytes allowed. Actual size (without debug info): {rpc_class_size} bytes.
-                        "#, contract_stem.clone()});
-                    }
-                }
                 artifact.artifacts.sierra = Some(file_name);
             }