--- a/scarb/src/compiler/mod.rs
+++ b/scarb/src/compiler/mod.rs
@@ -4,8 +4,10 @@ use cairo_lang_filesystem::ids::CrateId;
 pub use compilation_unit::*;
 pub use profile::*;
 pub use repository::*;
+use std::sync::mpsc;
 
 use crate::core::{TargetKind, Workspace};
+use crate::internal::artifacts_writer::Request;
 
 mod compilation_unit;
 mod compilers;
@@ -23,6 +25,7 @@ pub trait Compiler: Sync {
         &self,
         unit: &CairoCompilationUnit,
         cached_crates: &[CrateId],
+        artifacts_writer: mpsc::Sender<Request>,
         db: &mut RootDatabase,
         ws: &Workspace<'_>,
     ) -> Result<()>;