--- a/scarb/src/internal/mod.rs
+++ b/scarb/src/internal/mod.rs
@@ -1,3 +1,4 @@
+pub mod artifacts_writer;
 pub mod async_cache;
 pub mod cloneable_error;
 pub mod fsx;