… scarb/src/compiler/repository.rs
… scarb/src/internal/artifacts_writer.rs
… scarb/src/ops/compile.rs