Move serialization of .contract_class.json artifact to separate thread

Move serialization of .compiled_contract_class.json artifact to separate thread