@maciektr 2025-07-21
Before (last two PRs):
```
➜  cairo-contracts git:(v2.0.0) ✗ hyperfine --warmup 1 --runs 2 -- '/Users/maciektr/Projects/scarb/target/ci/scarb build --no-warnings --workspace --test'
Benchmark 1: /Users/maciektr/Projects/scarb/target/ci/scarb build --no-warnings --workspace --test
  Time (mean ± σ):     49.035 s ±  1.317 s    [User: 67.308 s, System: 5.224 s]
  Range (min … max):   48.103 s … 49.966 s    2 runs
```

After: 
```
➜  cairo-contracts git:(v2.0.0) ✗ hyperfine --warmup 1 --runs 2 -- '/Users/maciektr/Projects/scarb/target/ci/scarb build --no-warnings --workspace --test'
Benchmark 1: /Users/maciektr/Projects/scarb/target/ci/scarb build --no-warnings --workspace --test
  Time (mean ± σ):     43.716 s ±  0.157 s    [User: 67.853 s, System: 5.475 s]
  Range (min … max):   43.605 s … 43.828 s    2 runs
```